Extract Costs section from routes page
